feat: refactor authentication handling in Home and DentistCard components
