+++ src/components/DentistCard.tsx
@@ -1,7 +1,8 @@
 "use client";
 import { useRouter } from "next/navigation";
 import Image from "next/image";
-
+import Cookies from "js-cookie";
+import getMe from "@/libs/(auth)/getMe";
 interface DentistCardProps {
   name: string;
   expertise: string;
@@ -16,7 +17,15 @@ export const DentistCard: React.FC<DentistCardProps> = ({
   imageUrl = "/placeholder-dentist.jpg",
 }) => {
   const router = useRouter();
-
+  const token = Cookies.get("token");
+  const userCheck = async () => {
+    try {
+      await getMe(token as string);
+    } catch (err) {
+      Cookies.remove("token");
+    }
+  };
+  userCheck();
   return (
     <div className="relative p-6 bg-sky-200 rounded-3xl h-[500px] max-md:p-5 max-md:h-auto max-sm:p-4">
       <div className="text-3xl text-center">{name}</div>
@@ -35,7 +44,9 @@ export const DentistCard: React.FC<DentistCardProps> = ({
         {experience} Years of Experience
       </div>
       <button
-        onClick={() => router.push("/booking")}
+        onClick={() => {
+          token ? router.push("/booking") : router.push("/login");
+        }}
         className="absolute right-8 bottom-8 text-3xl bg-white rounded-3xl h-[85px] w-[146px] max-md:relative max-md:right-auto max-md:bottom-auto max-md:mx-auto max-md:my-5 max-sm:w-full max-sm:text-2xl max-sm:h-[60px] hover:bg-gray-100 transition-colors flex items-center justify-center"
       >
         Booking
